file helper methods moved to new class
